Compatibility and general improvements
init.lua: Enums will only be patched in if they don't already exists. For compatibility with other libraries.
init.lua: If an event that needs to be patched already has a listener, both listeners will be wrapped in a function. To ensure compatibility with other libraries.
init.lua: Changed the order of patches to make more sense, as well as returning a table value that contains the interaction object.
enum.lua: Removed buttonStyle since it is not related to interactions.
enum.lua: Added interactionCallbackType enum to for use internally.
Interaction.lua: Doc comment improvments and using enums where possible.